MessengerService/requirements.txt
Flask
SQLAlchemy
pymysql
requests
python-dotenv
APScheduler
alembic         # optional, if you want migrations
